Issue# 472: Added specification tests for WSN Connector.
