+++ transport/wsn/src/test/java/org/kaazing/gateway/transport/wsn/specification/ws/connector/ClosingHandshakeIT.java
@@ -1,0 +1,137 @@
+/*
+ * Copyright 2014, Kaazing Corporation. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kaazing.gateway.transport.wsn.specification.ws.connector;
+
+import static org.junit.Assert.assertTrue;
+import static org.kaazing.test.util.ITUtil.createRuleChain;
+
+import org.apache.mina.core.future.ConnectFuture;
+import org.apache.mina.core.service.IoHandler;
+import org.junit.Ignore;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TestRule;
+import org.kaazing.gateway.transport.IoHandlerAdapter;
+import org.kaazing.k3po.junit.annotation.Specification;
+import org.kaazing.k3po.junit.rules.K3poRule;
+import org.kaazing.mina.core.session.IoSessionEx;
+
+public class ClosingHandshakeIT {
+    private final K3poRule k3po = new K3poRule().setScriptRoot("org/kaazing/specification/ws/closing");
+
+    private final WsnConnectorRule connectorRule = new WsnConnectorRule();
+
+    @Rule
+    public TestRule chain = createRuleChain(connectorRule, k3po);
+
+
+    @Test
+    @Specification({
+        "server.send.empty.close.frame/handshake.response.and.frame" })
+    public void shouldCompleteCloseHandshakeWhenServerSendEmptyCloseFrame() throws Exception {
+        IoHandler handler = new IoHandlerAdapter<IoSessionEx>();
+        ConnectFuture connectFuture = connectorRule.connect("ws://localhost:8080/echo", null, handler);
+        connectFuture.awaitUninterruptibly();
+        assertTrue(connectFuture.isConnected());
+
+        k3po.finish();
+    }
+
+    @Test
+    @Specification({
+        "server.send.close.frame.with.code.1000/handshake.response.and.frame" })
+    public void shouldCompleteCloseHandshakeWhenServerSendCloseFrameWithCode1000() throws Exception {
+        IoHandler handler = new IoHandlerAdapter<IoSessionEx>();
+        ConnectFuture connectFuture = connectorRule.connect("ws://localhost:8080/echo", null, handler);
+        connectFuture.awaitUninterruptibly();
+        assertTrue(connectFuture.isConnected());
+
+        k3po.finish();
+    }
+
+    @Test
+    @Specification({
+        "server.send.close.frame.with.code.1000.and.reason/handshake.response.and.frame" })
+    public void shouldCompleteCloseHandshakeWhenServerSendCloseFrameWithCode1000AndReason() throws Exception {
+        IoHandler handler = new IoHandlerAdapter<IoSessionEx>();
+        ConnectFuture connectFuture = connectorRule.connect("ws://localhost:8080/echo", null, handler);
+        connectFuture.awaitUninterruptibly();
+        assertTrue(connectFuture.isConnected());
+
+        k3po.finish();
+    }
+
+    @Test
+    @Ignore("Gateway sends back the invalid UTF8 payload/reason instead of just sending the CLOSE code 1002")
+    @Specification({
+        "server.send.close.frame.with.code.1000.and.invalid.utf8.reason/handshake.response.and.frame" })
+    public void shouldFailWebSocketConnectionWhenServerSendCloseFrameWithCode1000AndInvalidUTF8Reason() throws Exception {
+        IoHandler handler = new IoHandlerAdapter<IoSessionEx>();
+        ConnectFuture connectFuture = connectorRule.connect("ws://localhost:8080/echo", null, handler);
+        connectFuture.awaitUninterruptibly();
+        assertTrue(connectFuture.isConnected());
+
+        k3po.finish();
+    }
+
+    @Test
+    @Specification({
+        "server.send.close.frame.with.code.1001/handshake.response.and.frame" })
+    public void shouldCompleteCloseHandshakeWhenServerSendCloseFrameWithCode1001() throws Exception {
+        IoHandler handler = new IoHandlerAdapter<IoSessionEx>();
+        ConnectFuture connectFuture = connectorRule.connect("ws://localhost:8080/echo", null, handler);
+        connectFuture.awaitUninterruptibly();
+        assertTrue(connectFuture.isConnected());
+
+        k3po.finish();
+    }
+
+    @Test
+    @Specification({
+        "server.send.close.frame.with.code.1005/handshake.response.and.frame" })
+    public void shouldFailWebSocketConnectionWhenServerSendCloseFrameWithCode1005() throws Exception {
+        IoHandler handler = new IoHandlerAdapter<IoSessionEx>();
+        ConnectFuture connectFuture = connectorRule.connect("ws://localhost:8080/echo", null, handler);
+        connectFuture.awaitUninterruptibly();
+        assertTrue(connectFuture.isConnected());
+
+        k3po.finish();
+    }
+
+    @Test
+    @Specification({
+        "server.send.close.frame.with.code.1006/handshake.response.and.frame" })
+    public void shouldFailWebSocketConnectionWhenServerSendCloseFrameWithCode1006() throws Exception {
+        IoHandler handler = new IoHandlerAdapter<IoSessionEx>();
+        ConnectFuture connectFuture = connectorRule.connect("ws://localhost:8080/echo", null, handler);
+        connectFuture.awaitUninterruptibly();
+        assertTrue(connectFuture.isConnected());
+
+        k3po.finish();
+    }
+
+    @Test
+    @Specification({
+        "server.send.close.frame.with.code.1015/handshake.response.and.frame" })
+    public void shouldFailWebSocketConnectionWhenServerSendCloseFrameWithCode1015() throws Exception {
+        IoHandler handler = new IoHandlerAdapter<IoSessionEx>();
+        ConnectFuture connectFuture = connectorRule.connect("ws://localhost:8080/echo", null, handler);
+        connectFuture.awaitUninterruptibly();
+        assertTrue(connectFuture.isConnected());
+
+        k3po.finish();
+    }
+}
